day8/enterprise-chat-agent/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==11.0.3
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
aiofiles==23.2.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
structlog==23.2.0
tenacity==8.2.3
cryptography>=41.0.0
pyjwt==2.8.0
